make use of handle_claim_burn in wallet daemon cli
