+++ applications/tari_dan_wallet_cli/Cargo.toml
@@ -14,6 +14,7 @@ tari_dan_common_types = { path = "../../dan_layer/common_types" }
 tari_dan_engine = { path = "../../dan_layer/engine" }
 tari_engine_types = { path = "../../dan_layer/engine_types" }
 tari_utilities = { git = "https://github.com/tari-project/tari_utilities.git", tag = "v0.4.10" }
+tari_dan_wallet_daemon = { path = "../tari_dan_wallet_daemon" }
 tari_wallet_daemon_client = { path = "../../clients/wallet_daemon_client" }
 tari_template_lib = { path = "../../dan_layer/template_lib" }
 tari_transaction_manifest = { path = "../../dan_layer/transaction_manifest" }
